Refactor code to keep all network calls in one module
