+++ synthea-hiv/uploader/main_test.py
@@ -17,8 +17,6 @@ import tempfile
 import unittest
 
 import main
-import mock
-
 
 class MainTest(unittest.TestCase):
 
@@ -44,64 +42,3 @@ class MainTest(unittest.TestCase):
     temp_file.flush()
     bundle = main.convert_to_bundle(temp_file.name)
     self.assertEqual(bundle.file_name, temp_file.name)
-
-  def test_fetch_location(self):
-    mock_sink = mock.MagicMock()
-    mock_sink.get_resource.return_value = {
-        'entry': [{
-            'not_location': 'location'
-        }]
-    }
-    location = main.fetch_location(mock_sink)
-    self.assertEqual(
-        location, {'8d6c993e-c2cc-11de-8d13-0010c6dffd0f': 'Unknown Location'})
-
-    mock_sink.get_resource.return_value = {
-        'entry': [{
-            'resource': {
-                'resourceType': 'Location',
-                'id': '8d6c993e-c2cc-11de-8d13-0010c6dffd0f',
-                'text': {
-                    'status': 'generated',
-                    'div':
-                        '<div xmlns="http://www.w3.org/1999/xhtml"><h2>Unknown '
-                        'Location</h2><h2/></div>'
-                },
-                'status': 'active',
-                'name': 'Unknown Location'
-            }
-        }, {
-            'resource': {
-                'resourceType': 'Location',
-                'id': '7f65d926-57d6-4402-ae10-a5b3bcbf7986',
-                'text': {
-                    'status': 'generated',
-                    'div':
-                        '<div '
-                        'xmlns="http://www.w3.org/1999/xhtml"><h2>Pharmacy</h2></div>'
-                },
-                'status': 'active',
-                'name': 'Pharmacy'
-            }
-        }, {
-            'resource': {
-                'resourceType': 'Location',
-                'id': '7fdfa2cb-bc95-405a-88c6-32b7673c0453',
-                'text': {
-                    'status': 'generated',
-                    'div':
-                        '<div '
-                        'xmlns="http://www.w3.org/1999/xhtml"><h2>Laboratory</h2></div>'
-                },
-                'status': 'active',
-                'name': 'Laboratory'
-            }
-        }]
-    }
-    location = main.fetch_location(mock_sink)
-    self.assertEqual(
-        location, {
-            '8d6c993e-c2cc-11de-8d13-0010c6dffd0f': 'Unknown Location',
-            '7f65d926-57d6-4402-ae10-a5b3bcbf7986': 'Pharmacy',
-            '7fdfa2cb-bc95-405a-88c6-32b7673c0453': 'Laboratory'
-        })
